[YS-425] [YS-424] refactor: 파일명 수정
